Move Authenticator and Compression to scylla-cql-core
